Reestruturação do código -Definir Afinidade ok
